Exclude methods defined in traits from export
Trait definitions are already excluded from the export and their methods
are flattened into the using classes. Exporting trait methods seperately
breaks the import in VW, as the corresponding class-id is undefined.

This commit
a) excludes trait methods and initializers and
b) changes the filter to #isClass instead of #isTrait for consistency.